simplify things and just use kube-apiserver for now
